+++ src/system/kernel/test.cpp
@@ -1,3 +1,3 @@
 #include <test.h>
 
-int64 _user_hello(int64 value) { return value; }
+int _user_hello(int value) { return value; }
